add qps attribute
