+++ last/llms/model/base_model.py
@@ -78,6 +78,9 @@ class HTTPAPILLMModel(BaseLLMModel):
         retry_options = ExponentialRetry(attempts = 2 ** 2, start_timeout = 1.0)
         # 同一个LLM的请求用同一个 session 访问
         self.retry_client = RetryClient(raise_for_status=False, retry_options=retry_options)
+        
+        # 不同平台 不同 QPS
+        self.qps = 4
     
     def __del__(self):
         loop = asyncio.get_running_loop()
